Split the event loop out of js-compute-builtins.cpp
